requirements.txt
streamlit == 1.21.0
pandas == 1.5.3
plotly == 5.15.0
seaborn == 0.12.2
streamlit-modal
streamlit-aggrid
folium
streamlit-folium
azure-core
azure-keyvault
azure-identity